+++ demo/CypressLayout.tsx
@@ -21,7 +21,9 @@ import {
   DropDownLink,
   RadioGroup,
   MultiSelect,
-  ToolTip
+  ToolTip,
+  TimePicker,
+  Validation
 } from './cypress';
 import { Styles } from './components/Styles';
 
@@ -85,6 +87,12 @@ export const CypressLayout = () => (
         </Route>
         <Route path="/cypress/tooltip">
           <ToolTip />
+        </Route>
+        <Route path="/cypress/timepicker">
+          <TimePicker />
+        </Route>
+        <Route path="/cypress/validation">
+          <Validation />
         </Route>
       </Switch>
     </L.Div>