+++ demo/cypress/index.ts
@@ -18,3 +18,5 @@ export { DropDownLink } from './DropDownLink';
 export { RadioGroup } from './RadioGroup';
 export { MultiSelect } from './MultiSelect';
 export { ToolTip } from './ToolTip';
+export { TimePicker } from './TimePicker';
+export { Validation } from './Validation';